Prefactor: Move `options` to separate file
